introduce BigBitSet on BitSet
git-svn-id: https://svn.ilb.ru/repos/common/trunk@167 a5b2bf95-3cba-4c8e-be42-a0ce6bbb2e66

+++ common-jpa/src/main/java/ru/ilb/common/jpa/bitset/BitAccessor.java
@@ -21,6 +21,6 @@ package ru.ilb.common.jpa.bitset;
  */
 public interface BitAccessor {
 
-    Long getBitNum(Object o);
+    int getBitNum(Object o);
     
 }
